third_party/blink/web_tests/external/wpt/media-source/last-frame-dimensions-expected.txt

This is a testharness.js-based test.
[FAIL] last frame dimensions
  assert_equals: videoWidth 640 expected 640 but got 320
Harness: the test ran to completion.
